Refactor assembly_strategies module
Separate classes of strategies to be in their own modules. Separate
related utilities to their own module.
